+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - Ability to use existing assembly for project via `odood assembly use` command.
   This could be useful in CI to automate assembly sync process.
 - Added new options to specify commit params for `odood assembly sync` command.
+- Added experimental `odood repo do-forward-port` command
 
 ### Changed
 - `odood log` now will automatically show the end of logfile